Add BasePayPalView and PayPalCreditView
