+++ src/html/main.html
@@ -36,7 +36,7 @@
           <div data-braintree-id="paypal-button" class="braintree-sheet__button--paypal"></div>
         </div>
       </div>
-      <div data-braintree-id="paypalCredit" class="braintree-paypal braintree-sheet">
+      <div data-braintree-id="paypalCredit" class="braintree-paypalCredit braintree-sheet">
         <div data-braintree-id="paypal-credit-sheet-header" class="braintree-sheet__header">
           <div class="braintree-sheet__header-label">
             <div class="braintree-sheet__logo--header">
