+++ src/scss/main.scss
@@ -702,7 +702,8 @@ $loader-scale-duration: 300ms;
 .braintree-show-card .braintree-card,
 .braintree-show-methods .braintree-methods,
 .braintree-show-options .braintree-options,
-.braintree-show-paypal .braintree-paypal {
+.braintree-show-paypal .braintree-paypal,
+.braintree-show-paypalCredit .braintree-paypalCredit {
   display: block;
   height: auto;
   overflow: visible;
