+++ src/views/payment-sheet-views/index.js
@@ -6,6 +6,6 @@ var result = {};
 
 result[paymentOptionIDs.card] = require('./card-view');
 result[paymentOptionIDs.paypal] = require('./paypal-view');
-result[paymentOptionIDs.paypalCredit] = require('./paypal-view');
+result[paymentOptionIDs.paypalCredit] = require('./paypal-credit-view');
 
 module.exports = result;
