+++ src/views/payment-sheet-views/paypal-credit-view.js
@@ -1,0 +1,16 @@
+'use strict';
+
+var paymentOptionIDs = require('../../constants').paymentOptionIDs;
+var BasePayPalView = require('./base-paypal-view');
+
+function PayPalCreditView() {
+  BasePayPalView.apply(this, arguments);
+
+  this._initialize(true);
+}
+
+PayPalCreditView.prototype = Object.create(BasePayPalView.prototype);
+PayPalCreditView.prototype.constructor = PayPalCreditView;
+PayPalCreditView.ID = PayPalCreditView.prototype.ID = paymentOptionIDs.paypalCredit;
+
+module.exports = PayPalCreditView;
